Solve second half of day 1 problem
The second half of the problem breaks the assumption that the character
to compare is the next character in the string.  So, the checksum
algorithm is modified to pick an arbitrary offset from the current
character and uses zip instead of sliding to create the pairs that need
to be compared.