Solve second half of Day 3
The change in the second half of Day 3 completely broke the initial
strategy, which relied on the fact that each value could easily be
located on the grid.  This was due to the fact that the numbers in each
ring were in a distinct range.

However, the new challenge breaks that and the values in each ring are
all over the place and depend on a specific progression through the
grid.

The new solution forces a walk through the grid, and Scala Streams were
used to setup the path without evaluating any more than necessary.