Solve second half of Day 6
The second half of this problem doesn't really require and new
processing, but a little bit more information about the processing that
is already being done.

Instead of simply returning the index of the duplicate memory state,
we need to return the index of both of the memory states that are
duplicated, which we can easily do by keeping a Map instead of a Set.