Solve second half of Day 7
The first half of Day 7 ignored the weights entirely and we simply
needed to find the root node, which was easy enough because it's the
only one that isn't the child of another node.

The second half requires you to actually build the tree and calculate
weights, so the solution build on the initial one but adds completely
new functionality.

I'm not particularly happy with it, but i've spent too much time on it,
so we're going to move on.