Solve second half of Day 8
The second half of this problem followed a similar pattern where the
main logic stayed the same, we just need to calculate a bit more
information about the processing.

Instead of just returning a tuple that includes the extra information, I
tried a different pattern where the code doing the work doesn't know
what the extra information is that's needed.  Instead, a "monitor" is
provided that wraps the main logic and the details of what is being
monitored is abstracted there.