Solve second half of Day 9
The original problem lended itself well to a visitor kind of pattern
where data was collected as the stream of text was being processed.

For the second half, since more information about the processing is
needed, a more formal abstraction for the visitor was added so that the
processing logic does as little as possible.