Solve second half of Day 17
The second half of this problem could have been solved using the
solution for the first half, but it wouldn't finish in a reasonable
amount of time.

In this case, we can optimize the solution by recognizing that the
desired value is always in the second position of the array.  This
allows us to just keep track of the last value that makes it into this
index without actually having to build the list itself.